fixbug  update actual time and style template report

+++ app/Http/Controllers/BackLogController.php
@@ -148,7 +148,7 @@ class BackLogController extends Controller
                     foreach ($activity->content->changes as $change) {
                         if ($change->field == 'actualHours') {
                             $projectList[$activity->project->name][$pi]['task']        = $activity->content->summary;
-                            $projectList[$activity->project->name][$pi]['actualHours'] = (int)$change->new_value - (int)$change->old_value;
+                            $projectList[$activity->project->name][$pi]['actualHours'] = (float)$change->new_value - (float)$change->old_value;
                         }
                     }
